+++ exspy/_misc/eds/__init__.pyi
@@ -1,46 +1,6 @@
 from .ffast_mac import ffast_mac, ffast_mac_db
-from .utils import (
-    _get_element_and_line,
-    _get_energy_xray_line,
-    _parse_only_lines,
-    cross_section_to_zeta,
-    electron_range,
-    get_abs_corr_cross_section,
-    get_abs_corr_zeta,
-    get_FWHM_at_Energy,
-    get_xray_lines,
-    get_xray_lines_near_energy,
-    print_lines,
-    print_lines_near_energy,
-    quantification_cliff_lorimer,
-    quantification_cross_section,
-    quantification_zeta_factor,
-    take_off_angle,
-    xray_lines_model,
-    xray_range,
-    zeta_to_cross_section,
-)
 
 __all__ = [
-    "_get_element_and_line",
-    "_get_energy_xray_line",
-    "_parse_only_lines",
-    "cross_section_to_zeta",
-    "electron_range",
     "ffast_mac",
     "ffast_mac_db",
-    "get_abs_corr_cross_section",
-    "get_abs_corr_zeta",
-    "get_FWHM_at_Energy",
-    "get_xray_lines",
-    "get_xray_lines_near_energy",
-    "print_lines",
-    "print_lines_near_energy",
-    "quantification_cliff_lorimer",
-    "quantification_cross_section",
-    "quantification_zeta_factor",
-    "take_off_angle",
-    "xray_lines_model",
-    "xray_range",
-    "zeta_to_cross_section",
 ]
