+++ exspy/_misc/eels/__init__.pyi
@@ -2,14 +2,10 @@ from .gosh_gos import GoshGOS
 from .gosh_gos_source import GOSH_SOURCES
 from .hydrogenic_gos import HydrogenicGOS
 from .hartree_slater_gos import HartreeSlaterGOS
-from .effective_angle import effective_angle
-from .dielectric import eels_constant_dielectric
 
 __all__ = [
     "GOSH_SOURCES",
     "HydrogenicGOS",
     "GoshGOS",
     "HartreeSlaterGOS",
-    "effective_angle",
-    "eels_constant_dielectric",
 ]
