--- exspy/utils/eds.py
@@ -1,47 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2007-2025 The eXSpy developers
-#
-# This file is part of eXSpy.
-#
-# eXSpy is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# eXSpy is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with eXSpy. If not, see <https://www.gnu.org/licenses/#GPL>.
-
-
-from exspy._misc.eds.utils import (
-    cross_section_to_zeta,
-    electron_range,
-    get_xray_lines,
-    get_xray_lines_near_energy,
-    print_lines,
-    print_lines_near_energy,
-    take_off_angle,
-    xray_range,
-    zeta_to_cross_section,
-)
-
-
-__all__ = [
-    "cross_section_to_zeta",
-    "electron_range",
-    "get_xray_lines",
-    "get_xray_lines_near_energy",
-    "print_lines",
-    "print_lines_near_energy",
-    "take_off_angle",
-    "xray_range",
-    "zeta_to_cross_section",
-]
-
-
-def __dir__():
-    return sorted(__all__)
