+++ exspy/utils/eds/__init__.pyi
@@ -1,0 +1,46 @@
+from ._geometry import take_off_angle
+from ._quantification import (
+    cross_section_to_zeta,
+    get_abs_corr_cross_section,
+    get_abs_corr_zeta,
+    quantification_cliff_lorimer,
+    quantification_cross_section,
+    quantification_zeta_factor,
+    zeta_to_cross_section,
+)
+from ._particle_matter_interaction import electron_range, xray_range
+from ._xray_lines import (
+    _get_element_and_line,
+    _get_energy_xray_line,
+    _get_xray_lines_family,
+    _parse_only_lines,
+    get_FWHM_at_Energy,
+    get_xray_lines,
+    get_xray_lines_near_energy,
+    print_lines,
+    print_lines_near_energy,
+    xray_lines_model,
+)
+
+__all__ = [
+    "_get_element_and_line",
+    "_get_energy_xray_line",
+    "_get_xray_lines_family",
+    "_parse_only_lines",
+    "cross_section_to_zeta",
+    "electron_range",
+    "get_abs_corr_cross_section",
+    "get_abs_corr_zeta",
+    "get_FWHM_at_Energy",
+    "get_xray_lines",
+    "get_xray_lines_near_energy",
+    "print_lines",
+    "print_lines_near_energy",
+    "quantification_cliff_lorimer",
+    "quantification_cross_section",
+    "quantification_zeta_factor",
+    "take_off_angle",
+    "xray_lines_model",
+    "xray_range",
+    "zeta_to_cross_section",
+]
