+++ exspy/utils/eels/__init__.pyi
@@ -1,0 +1,18 @@
+from .dielectric import eels_constant_dielectric
+from .edges import get_edges_near_energy, get_info_from_edges
+from .effective_angle import effective_angle
+from .electron_inelastic_mean_free_path import (
+    iMFP_Iakoubovskii,
+    iMFP_TPP2M,
+    iMFP_angular_correction,
+)
+
+__all__ = [
+    "eels_constant_dielectric",
+    "effective_angle",
+    "get_edges_near_energy",
+    "get_info_from_edges",
+    "iMFP_angular_correction",
+    "iMFP_Iakoubovskii",
+    "iMFP_TPP2M",
+]
